pyproject.toml
[project]
name = "doc-flow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.11",
    "langgraph>=1.0.10",
    "langgraph-cli[inmem]>=0.4.15",
    "pydantic-settings>=2.7.0",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=0.25.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
